Rework the font configuration
Rather than specifying specific fonts for each operating system / environment,
specify a list of fonts in order of preference, so it uses the best one
available given the current environment.  It automatically adjusts the
fontstrings as necessary given the requirements of the OS.

This also results in Inconsolata being preferred over Monaco (the previous OSX
font setting).

Signed-off-by: Chris Larson <chris_larson@mentor.com>