Enable the 'S' keymap for git rebase -i
This was mentioned in a comment in gitrebase.vim.  It binds 'S' to a :Cycle
command when in a rebase -i, which cycles through the available actions for a
given commit.  Mouse over the 'pick' and each time you hit S it moves to the
next available action (i.e. edit, squash, fixup).

Signed-off-by: Chris Larson <chris_larson@mentor.com>